web1/web-captcha/requirements.txt
Flask==1.1.2
pycryptodome==3.9.9
Pillow==9.0.1
itsdangerous==1.1.0